Force-retry jobs when connection is lost
We already reconnect when the PG connection is lost mid-job, but the job
would normally get re-enqueued on the failed_jobs queue which isn't
ideal. Jobs should almost always be retried if the connection is lost. I
added a `disable_retries!` escape hatch for the rare occasion that
they're not.

See rainforestapp/Rainforest#7995