RF-29953 Change QueueClassicPlus::Base to enqueue without can_enqueue?
The enqueueability is pushed down to the queue level with the uniqueness
constraint, but we have to pass in if the job is locked?.

Here we have to forego the QUEUE_CLASSIC_MAX_LOCK_TIME feature because there
is no way to pass a time based condition to the constraint at runtime.
An alternative might be to hard code the max lock time in the constraint.

But I think having a lock "expire" is surprising behaviour, almost never
accounted for, and can cause bugs if not handled. So removing this feature.